Latest changes. PR ready.
